WIP Tabs moved to their own pages, condition and power saving/loading not working currently.
