+++ advantages.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>M&M Useful Tools</title>
+</head>
+<body>
+    <div id="navbar-container"></div>
+    <div class="container">
+        <h2>Advantages List</h2>
+        <div class="search-container">
+            <input type="text" id="searchAdvantages" placeholder="Enter advantage name...">
+            <label for="filterAdvantagesCheckbox">
+                <input type="checkbox" id="filterAdvantagesCheckbox">
+                Filter Marked Advantages
+            </label>
+        </div>
+        <ul id="advantagesList"></ul>
+    </div>
+    <!--TODO: CLEAN THIS UP-->
+    <div id="tooltipPopup" class="tooltip-popup">
+        <div class="tooltip-content"><strong id="tooltipTitle" class="tooltipTitle"></strong><p id="tooltipContent"></p></div>
+    </div>
+</body>
+</html>
