+++ conditions.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>M&M Useful Tools</title>
+</head>
+<body>
+    <div id="navbar-container"></div>
+    <div class="container container-wide">
+        <h2>Conditions</h2>
+        <div class="split">
+            <div class="half">
+                <h3>Base Conditions</h3>
+                <ul id="baseConditionsUList" class="conditionsList"></ul>
+                <h3>Combined Conditions</h3>
+                <ul id="combinedConditionsUList" class="conditionsList"></ul>
+            </div>
+            <div class="half">
+                <button id="addCreatureBtn">Add Creature</button>
+                <ul id="creaturesList"></ul>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
