+++ index.html
@@ -22,83 +22,6 @@
                 <div class="result" id="result"></div>
             </div>
         </div>
-        <div id="measurementsTab" class="tab">
-            <div class="container">
-                <h2>Measurements Table</h2>
-                <table>
-                    <thead>
-                        <tr>
-                            <th colspan="5" style="background-color: transparent; border: 1px solid transparent; position: relative; ">
-                                <button onclick="populateMeasurementsTable()" id="toggleButton" value="imperial">Toggle</button>
-                            </th>
-                        </tr>
-                        <tr>
-                            <th>Rank</th>
-                            <th>Mass</th>
-                            <th>Time</th>
-                            <th>Distance</th>
-                            <th>Volume</th>
-                        </tr>
-                    </thead>
-                    <tbody id="measurementsTableBody">
-                    </tbody>
-                </table>
-            </div>
-        </div>
-        <div id="advantagesTab" class="tab">
-            <div class="container">
-                <h2>Advantages List</h2>
-                <div class="search-container">
-                    <input type="text" id="searchAdvantages" placeholder="Enter advantage name...">
-                    <label for="filterAdvantagesCheckbox">
-                        <input type="checkbox" id="filterAdvantagesCheckbox">
-                        Filter Marked Advantages
-                    </label>
-                </div>
-                <ul id="advantagesList"></ul>
-            </div>
-            <!--TODO: CLEAN THIS UP-->
-            <div id="tooltipPopup" class="tooltip-popup">
-                <div class="tooltip-content"><strong id="tooltipTitle" class="tooltipTitle"></strong><p id="tooltipContent"></p></div>
-            </div>
-        </div>
-        <div id="conditionsTab" class="tab">
-            <div class="container container-wide">
-                <h2>Conditions</h2>
-                <div class="split">
-                    <div class="half">
-                        <h3>Base Conditions</h3>
-                        <ul id="baseConditionsUList" class="conditionsList"></ul>
-                        <h3>Combined Conditions</h3>
-                        <ul id="combinedConditionsUList" class="conditionsList"></ul>
-                    </div>
-                    <div class="half">
-                        <button id="addCreatureBtn">Add Creature</button>
-                        <ul id="creaturesList"></ul>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div id="powersTab" class="tab">
-            <div class="container container-max">
-                <h2>Powers</h2>
-                <div>
-                    <label>Total Active Power Point Cost: <span id="totalPointCost">0</span></label>
-                </div>
-                <div>
-                    <label for="powersNumColumns">Enter number of columns:</label>
-                    <input type="number" id="powersNumColumns" placeholder="Enter number of columns">
-                    <button onclick="generatePowerTable()">Generate Table</button>
-
-                    <button id="importPowersBtn">Import Powers</button>
-                    <button id="exportPowersBtn">Export Powers</button>
-                    <input type="file" id="importPowers" accept=".json" style="display: none;">
-                </div>
-                <button id="addPowerBtn">Add Power</button>
-
-                <div id="powersTableContainer"></div>
-            </div>
-        </div>
     </div>
 </body>
 </html>
