+++ measurements-table.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>M&M Useful Tools</title>
+</head>
+<body>
+    <div id="navbar-container"></div>
+    <div class="container">
+        <h2>Measurements Table</h2>
+        <table>
+            <thead>
+                <tr>
+                    <th colspan="5" style="background-color: transparent; border: 1px solid transparent; position: relative; ">
+                        <button id="toggleButton" value="imperial">Toggle</button>
+                    </th>
+                </tr>
+                <tr>
+                    <th>Rank</th>
+                    <th>Mass</th>
+                    <th>Time</th>
+                    <th>Distance</th>
+                    <th>Volume</th>
+                </tr>
+            </thead>
+            <tbody id="measurementsTableBody">
+            </tbody>
+        </table>
+    </div>
+</body>
+</html>
