+++ navbar.html
@@ -2,11 +2,11 @@
     <h1>Mutants and Masterminds Useful Tools</h1>
 
     <div class="topnav">
-        <button id="degreeCalcTabNavBtn">Degrees Calculator</button>
-        <a href="throwing-distance-calculator.html"><button id="throwingDistCalcTabNavBtn">Throwing Distance Calculator</button></a>
-        <button id="measurementsTabNavBtn">Measurements Table</button>
-        <button id="advantagesTabNavBtn">Advantages</button>
-        <button id="conditionsTabNavBtn">Conditions</button>
-        <button id="powersTabNavBtn">Variable Powers Manager</button>
+        <a href="index.html" class="button-link">Degrees Calculator</a>
+        <a href="throwing-distance-calculator.html" class="button-link">Throwing Distance Calculator</a>
+        <a href="measurements-table.html" class="button-link">Measurements Table</a>
+        <a href="advantages.html" class="button-link">Advantages</a>
+        <a href="conditions.html" class="button-link">Conditions</a>
+        <a href="variable-powers-manager.html" class="button-link">Variable Powers Manager</a>
     </div>
 </div>
