+++ variable-powers-manager.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>M&M Useful Tools</title>
+</head>
+<body>
+    <div id="navbar-container"></div>
+    <div class="container container-max">
+        <h2>Powers</h2>
+        <div>
+            <label>Total Active Power Point Cost: <span id="totalPointCost">0</span></label>
+        </div>
+        <div>
+            <label for="powersNumColumns">Enter number of columns:</label>
+            <input type="number" id="powersNumColumns" placeholder="Enter number of columns">
+            <button onclick="generatePowerTable()">Generate Table</button>
+
+            <button id="importPowersBtn">Import Powers</button>
+            <button id="exportPowersBtn">Export Powers</button>
+            <input type="file" id="importPowers" accept=".json" style="display: none;">
+        </div>
+        <button id="addPowerBtn">Add Power</button>
+
+        <div id="powersTableContainer"></div>
+    </div>
+</body>
+</html>
